requirements.txt
# START OF FILE requirements.txt #
python-telegram-bot>=20.0
pydub>=0.25.1
openai>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
langchain-community>=0.2.0 # ИЗМЕНЕНИЕ: Обновляем
pypdf>=4.0.0
python-docx>=1.1.2
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
# ИЗМЕНЕНИЕ: Обновляем ctransformers до последней версии для поддержки Llama-3.2
ctransformers[cuda]>=0.2.27

pywhispercpp>=1.1.2

# Зависимости для аудио и numpy
librosa>=0.10.0
# Жестко фиксируем версию numpy для избежания проблем совместимости
numpy<2.0

# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
# END OF FILE requirements.txt #